Did scroll-snap-align. The scroll-snap-align property in CSS defines the alignment point on a child element where the snapping should occur relative to its parent scroll container's viewport.
Essentially, it tells the browser where on the target element should "stick" to the snap container's edge. Use the snap-center utility to snap an element to its center when being scrolled inside a snap container, Use the snap-start utility to snap an element to its start when being scrolled inside a snap container and Use the snap-end utility to snap an element to its end when being scrolled inside a snap container.